# libs/api-client/src/lib/request-types.ts
import type { components } from './types.generated';

export type CreateCategoryDto = components['schemas']['CreateCategoryDto'];
export type UpdateCategoryDto = components['schemas']['UpdateCategoryDto'];

export type CreateMenuItemDto = components['schemas']['CreateMenuItemDto'];
export type UpdateMenuItemDto = components['schemas']['UpdateMenuItemDto'];

export type CreateTableDto = components['schemas']['CreateTableDto'];
export type UpdateTableDto = components['schemas']['UpdateTableDto'];
export type UpdateTableStatusDto = components['schemas']['UpdateTableStatusDto'];

export type CreateOrderDto = components['schemas']['CreateOrderDto'];
export type OrderItemDto = components['schemas']['OrderItemDto'];
export type AddOrderItemsDto = components['schemas']['AddOrderItemsDto'];
export type UpdateOrderItemStatusDto = components['schemas']['UpdateOrderItemStatusDto'];

export type PaymentAllocationDto = components['schemas']['PaymentAllocationDto'];
export type CreatePaymentDto = components['schemas']['CreatePaymentDto'];